sam_baseline2/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
opencv-python>=4.5.0
numpy>=1.20.0
Pillow>=8.0.0
matplotlib>=3.3.0
tqdm>=4.60.0
